chore: fix PhoenixStorybook search by avoiding double LiveSocket
Point js_path at a thin storybook.js that only sets window.storybook.Hooks;
PhoenixStorybook’s bundle owns LiveSocket.connect(). Loading app.js first
caused two sockets and broke quick search and related UI.
